+++ src/routes/Layout/index.tsx
@@ -1,9 +1,9 @@
 import { Outlet } from 'react-router-dom'
 
 import styles from './layout.module.scss'
-import Sidebar from '../Sidebar'
-import Header from '../Header'
-import Breadcrumb from '../Breadcrumb'
+import Sidebar from './Sidebar'
+import Header from './Header'
+import Breadcrumb from './Breadcrumb'
 
 const Layout = (): JSX.Element => {
   return (
